+++ src/components/ui/index.ts
@@ -1,8 +1,10 @@
 // UI component exports
-// Placeholder files will be implemented in future tasks
-// export * from './Button'
-// export * from './Card'
-// export * from './Badge'
-// export * from './Modal'
-// export * from './Tooltip'
-// export * from './ScrollIndicator'
+export * from './Button'
+export * from './Container'
+export * from './Section'
+export * from './Heading'
+export * from './Card'
+export * from './Badge'
+export * from './Modal'
+export * from './Tooltip'
+export * from './ScrollIndicator'