+++ src/lib/utils/index.ts
@@ -1,5 +1,4 @@
 // Utility exports
-// Placeholder files will be implemented in future tasks
-// export * from './cn'
+export * from './cn'
 // export * from './scroll'
 // export * from './validation'